+++ src/emby_cli/api/__init__.py
@@ -2,5 +2,6 @@
 
 from emby_cli.api.collections import CollectionsService
 from emby_cli.api.items import ItemsService
+from emby_cli.api.libraries import LibrariesService
 
-__all__ = ["CollectionsService", "ItemsService"]
+__all__ = ["CollectionsService", "ItemsService", "LibrariesService"]